refactor: 타입 분리 적용
